refactor(client_utils): refactor data generator function
This refactor adjusts the `client_utils.py` module in response to previous changes in the data formatting logic, specifically the restructuring of `TrainDataset` and `load_train_data` in `cyberid_pkg.dataformat.py`.

I could have only changed the logic of the Dataset class, by adding an indexing function encompassing the old concept, but that would lead to 2 unexpected points:
1- having the modifications in a Dataset class does not affect the speed of processing.
2- load_train_data is a dynamic state function, which makes the generation process untrackable.
Which led to radically changing the class.

### Changes:
- Extracted and adapted the data generator function to align with the fixed-state implementation introduced in the prior commit.
- Ensured compatibility with the new `TrainDataset` class structure and the modified `load_train_data()` function.

### Why this is important:
This refactor allows `client_utils.py` to operate efficiently with the newly updated data formatting system, by making a global state of the data only once, for the whole process. It addresses dependencies affected by the recent overhaul, improving data handling consistency and reproducibility.